modules/location-service/requirements.txt
kafka-python==2.0.2
grpcio==1.47.0
protobuf==3.19.4